Create mersenne_numbers.java
A Mersenne prime is a prime number that can be written in the form \(2^{n}-1\). For example \(31\) is a Mersenne prime that can be written as \(2^{5}-1\). The first few Mersenne primes are \(3, 7, 31, 127, 8191\). There are 50 known Mersenne primes as of June 2018, though we hope that it will change in the future. An interesting thing about Mersenne primes is that
they are the easiest natural numbers to prove to be primes, so they make up the largest category on the list of known prime numbers